Fix I/O in test
