+++ tests/components/tradfri/test_config_flow.py
@@ -18,14 +18,6 @@ def mock_auth():
 
 
 @pytest.fixture
-def mock_gateway_info():
-    """Mock get_gateway_info."""
-    with patch('homeassistant.components.tradfri.config_flow.'
-               'get_gateway_info') as mock_gateway:
-        yield mock_gateway
-
-
-@pytest.fixture
 def mock_entry_setup():
     """Mock entry setup."""
     with patch('homeassistant.components.tradfri.'
